requirements.txt
pycord
requests
asyncio
asyncssh
pyyaml